+++ api/application/use_cases/card_use_cases.py
@@ -4,7 +4,7 @@ from uuid import UUID, uuid4
 from api.application.dtos.card_dto import CardCreate, CardResponse, CardUpdate
 from api.domain.entities.bank import Bank
 from api.domain.entities.card import Card
-from api.domain.exceptions import NotFoundError
+from api.domain.exceptions import ForbiddenError, NotFoundError
 from api.domain.repositories.bank_repository import BankRepository
 from api.domain.repositories.card_repository import CardRepository
 
@@ -13,8 +13,8 @@ class ListCardsUseCase:
     def __init__(self, repo: CardRepository):
         self._repo = repo
 
-    async def execute(self, user_id: UUID) -> list[CardResponse]:
-        cards = await self._repo.list_by_user(user_id)
+    async def execute(self, household_id: UUID) -> list[CardResponse]:
+        cards = await self._repo.list_by_household(household_id)
         return [CardResponse.model_validate(c, from_attributes=True) for c in cards]
 
 
@@ -22,10 +22,12 @@ class GetCardUseCase:
     def __init__(self, repo: CardRepository):
         self._repo = repo
 
-    async def execute(self, card_id: UUID) -> CardResponse:
+    async def execute(self, card_id: UUID, household_id: UUID) -> CardResponse:
         card = await self._repo.get_by_id(card_id)
         if card is None:
             raise NotFoundError("Card", str(card_id))
+        if card.household_id != household_id:
+            raise ForbiddenError("Acesso negado a este recurso")
         return CardResponse.model_validate(card, from_attributes=True)
 
 
@@ -34,7 +36,9 @@ class CreateCardUseCase:
         self._card_repo = card_repo
         self._bank_repo = bank_repo
 
-    async def execute(self, data: CardCreate, user_id: UUID) -> CardResponse:
+    async def execute(
+        self, data: CardCreate, user_id: UUID, household_id: UUID
+    ) -> CardResponse:
         bank_id = data.bank_id
         if bank_id is None:
             bank = Bank(
@@ -43,6 +47,7 @@ class CreateCardUseCase:
                 color="#6B7280",
                 is_default=False,
                 user_id=user_id,
+                household_id=household_id,
             )
             created_bank = await self._bank_repo.create(bank)
             bank_id = created_bank.id
@@ -51,6 +56,7 @@ class CreateCardUseCase:
             id=uuid4(),
             user_id=user_id,
             bank_id=bank_id,
+            household_id=household_id,
             name=data.name,
             type=data.type,
             last_digits=data.last_digits,
@@ -67,10 +73,14 @@ class UpdateCardUseCase:
     def __init__(self, repo: CardRepository):
         self._repo = repo
 
-    async def execute(self, card_id: UUID, data: CardUpdate) -> CardResponse:
+    async def execute(
+        self, card_id: UUID, data: CardUpdate, household_id: UUID
+    ) -> CardResponse:
         card = await self._repo.get_by_id(card_id)
         if card is None:
             raise NotFoundError("Card", str(card_id))
+        if card.household_id != household_id:
+            raise ForbiddenError("Acesso negado a este recurso")
         update_data = data.model_dump(exclude_unset=True)
         updated_card = card.model_copy(update=update_data)
         updated = await self._repo.update(updated_card)
@@ -81,8 +91,10 @@ class DeleteCardUseCase:
     def __init__(self, repo: CardRepository):
         self._repo = repo
 
-    async def execute(self, card_id: UUID) -> None:
+    async def execute(self, card_id: UUID, household_id: UUID) -> None:
         card = await self._repo.get_by_id(card_id)
         if card is None:
             raise NotFoundError("Card", str(card_id))
+        if card.household_id != household_id:
+            raise ForbiddenError("Acesso negado a este recurso")
         await self._repo.delete(card_id)
